% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{script_comment}
\alias{script_comment}
\title{Comment Type to Script R}
\usage{
script_comment(x, 
 type_comment = 'h1',
 type_line = '-',
 type_size = NULL,
 clipboard = TRUE)
}
\arguments{
\item{x}{character class vector that indicates the message of the title or subtitle of the comment.}

\item{type_comment}{character class vector that indicates the comment type with option 'h1', 'h2' or 'h3'.}

\item{type_line}{character class vector that indicates the format type to header.}

\item{type_size}{numeric class vector that indicates the length to header.}

\item{clipboard}{logical class vector that indicates output to clipboard or console with option 'TRUE' or 'FALSE'.}
}
\value{
character class vector that indicates the comment type for R script.
}
\description{
Creates title and subtitle as comment to organize the R script.
}
\examples{
# default
# script_comment("test")

# other usage
# h1
script_comment('test_1', clipboard = FALSE)
script_comment(c("test_1", "test_2"), clipboard = FALSE)

# h2
script_comment('test_1', type_comment = "h2", clipboard = FALSE)
script_comment(c("test_1", "test_2"), type_comment = "h2", clipboard = FALSE)

# h3
script_comment('test_1', type_comment = "h3", clipboard = FALSE)
script_comment(c("test_1", "test_2"), 
 type_comment = "h3", type_size = 40, clipboard = FALSE)
}
