% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{script_count}
\alias{script_count}
\title{Count Line of Code}
\usage{
script_count(path, exclude = "#|^$")
}
\arguments{
\item{path}{character class vector that indicates the file path.}

\item{exclude}{character class vector that indicates the default to exclude with regular expression.}
}
\value{
character class vector that indicates the code line count for R script.
}
\description{
Count of code line written in script R. The default counting process ignores comment and empty.
}
\examples{
# default
# script_count(path)
}
