% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{script_create}
\alias{script_create}
\title{Create R Script}
\usage{
script_create(path, 
 name_script, 
 name_author = NULL, 
 name_project = NULL, 
 type_line = "-", 
 type_size = 42, 
 add_line = NULL, 
 rm_line = NULL)
}
\arguments{
\item{path}{character class vector that indicates the directory path.}

\item{name_script}{character class vector that indicates the file name.}

\item{name_author}{character class vector that indicates the author name.}

\item{name_project}{character class vector that indicates the author project.}

\item{type_line}{character class vector that indicates the format type to header.}

\item{type_size}{numeric class vector that indicates the length to header.}

\item{add_line}{numeric class vector that indicates the vector to add in the header.}

\item{rm_line}{numeric class vector that indicates the vector to remove in the header.}
}
\value{
the R script file with code development environment information in header.
}
\description{
Creates the R script with the information header of the code development environment, author, and more.
}
\examples{
# script_create(getwd(), "test 1")
# script_create(getwd(), c("test 1", "test 2"))
}
