% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{script_head}
\alias{script_head}
\title{Header to R Script}
\usage{
script_head(name_author = NULL, 
 name_project = NULL, 
 type_language = "r", 
 type_line = "-", 
 type_size = 42, 
 add_line = NULL, 
 rm_line = NULL, 
 clipboard = TRUE)
}
\arguments{
\item{name_author}{character class vector that indicates the author name.}

\item{name_project}{character class vector that indicates the author project.}

\item{type_language}{character class vector that indicates the language version with option 'r', 'python' and 'julia'.}

\item{type_line}{character class vector that indicates the format type to header.}

\item{type_size}{numeric class vector that indicates the length to header.}

\item{add_line}{numeric class vector that indicates the vector to add in the header.}

\item{rm_line}{numeric class vector that indicates the vector to remove in the header.}

\item{clipboard}{logical class vector that indicates output to clipboard or console with option 'TRUE' or 'FALSE'.}
}
\value{
character class vector with code development environment information.
}
\description{
Creates header with information from the code development environment, author, and more.
}
\examples{
# default
# script_head()

# usage other
script_head(clipboard = FALSE)
script_head(name_author = "test", clipboard = FALSE)
script_head(name_author = c("test 1", "test 2", "test 3"), clipboard = FALSE)
script_head(name_project = "test 2", clipboard = FALSE)
script_head(add_line = "Description: test", 
 rm_line = c("Platform", "GUI", "Author"), 
 clipboard = FALSE)
}
